pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "boto3>=1.38.27",
    "fastapi-cli>=0.0.7",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.20",
]
